undo fadeInOut layers, change border color
fadeInOut on the layers DID show an improvement in performance, which is
good. I'll consider this for implementation later on.

On another note, I noticed the border color is using rgba which could
also be a reason why things are getting rough when there's a lot of
layers. Gonna replace that with a solid color to see if there's an
improvement when all the layers are visible.